Ads work perfectly Results are displayed absolutely amazing AdReminder Dialog has been added Speed change has been added Custom Color has been added TempResultsFile has been added OnlyColorMode has been revised Made app to standard design SettingsButton in Results not working good on color Done!

+++ app/src/main/java/com/notorein/bt/Strings.java
@@ -357,7 +357,7 @@ public class Strings {
 //        add = Strings.speedText + Strings.changeIntervalInfoText + fadeInterval + Strings.changeIntervalInfoTextII;
         add = Strings.changeIntervalInfoText + fadeInterval
                 + Strings.changeIntervalInfoTextII + Strings.speedText + add
-                + +(percentageCorrected) + "%  "
+                + +((int)percentageCorrected) + "%  "
                 + Strings.estimatedTrialLengthText
                 + lengthTrial
                 + textViewTimeAdd_II
